register-server: Fix status update
Took status from object that already existed but had no status set
because status cannot be set on creation. This leads to `the object
has been modified` errors a few times until creation is successful,
leading to the creation of unnecessary Machine objects and secrets,
which was especially observable with the parallel attestation test.

Instead, try setting three times with a delay and warn if unsuccessful.

Signed-off-by: Jakob Naucke <jnaucke@redhat.com>